kipya/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.10
django-classy-tags==1.0.0
django-cms==3.7.1
django-cors-headers==3.2.1
django-filer==1.7.0
django-filter==2.2.0
django-formtools==2.2
django-haystack==2.8.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-newsletter==0.9
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.5.0
djangocms-column==1.10.0
djangocms-file==2.4.0
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.2.1
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
djangorestframework==3.11.0
easy-thumbnails==2.7
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
kombu==4.6.8
ldif3==3.1.1
Markdown==3.2.1
mysqlclient==1.4.6
Pillow==7.0.0
pkg-resources==0.0.0
PyPDF2==1.26.0
python-card-me==0.9.3
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.42
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
stripe==2.42.0
surlex==0.2.0
tzlocal==2.0.0
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
zipp==3.1.0
